requirements.txt
streamlit
requests
beautifulsoup4
flask
python-dotenv
